Initial Commit - Starting Repository to commit updates
Very first commit for the Fosternaut website. This website was an idea I had as I currently work as a volunteer for my city's local animal shelter. I foster cats and kittens and noticed that I have very little insight into the  shelter's daily practices or the outcomes of the foster's once I return them back for adoption. I know the shelter is overwhelmed with work and it's out of their scope to manage keeping the volunteer's informed, so I thought it would be a fun idea to create a functional website to help shorten that gap. My site aims to get the volunteers to all work together to help the shelter staff with scheduling, social networking, animal updates, posts and any other vital informational we all want.

Technical speak: I am currently working on building a full-stack site which will also contain access to a database using SQL. I am considering changing the HTML/CSS/Javascript to React once I learn more about the language.